Refactor SDK tests (without DHT) (#275)
